src/main/java/cn/edu/gzccc/sell/service/impl/CategoryServiceImpl.java
package cn.edu.gzccc.sell.service.impl;

import cn.edu.gzccc.sell.dataobject.ProductCategory;
import cn.edu.gzccc.sell.repository.ProductCategoryRepository;
import cn.edu.gzccc.sell.service.Categoryservice;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryServiceImpl implements Categoryservice {

    @Autowired
    private ProductCategoryRepository productCategoryRepository;

    @Override
    public ProductCategory findOne(Integer categoryid) {
        return productCategoryRepository.findOne(categoryid);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
